MB-6643 Swap PPM image on Move 101 page
Previously contained foreign service members.
